requirements.txt
plotly
streamlit
python-dotenv
requests
joblib
scikit-learn
numpy
pandas
